+++ src/domain/exercises/index.ts
@@ -2,3 +2,4 @@ export * from './generators';
 export * from './grading';
 export * from './speech';
 export * from './types';
+export * from './variations';